document_backend/app/database.py

# # app/database.py
# import os
# from sqlalchemy import create_engine
# from sqlalchemy.orm import sessionmaker, declarative_base
# from dotenv import load_dotenv

# load_dotenv()
# DATABASE_URL = os.getenv("DATABASE_URL")

# # SQLite database URL (file will be created locally)


# # SQLAlchemy engine
# engine = create_engine(
#     DATABASE_URL, connect_args={"check_same_thread": False}  # Required for SQLite
# )

# # Session factory
# SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

# # Base class for models
# Base = declarative_base()

# # FastAPI dependency to get DB session
# def get_db():
#     db = SessionLocal()
#     try:
#         yield db
#     finally:
#         db.close()

# # Initialize all tables
# # def init_db():
# #     import app.models.profile_model
# #     import app.models.document_model
# #     import app.models.summary_model
# #     import app.models.qa_model
# #     import app.models.embedding_model
# #     Base.metadata.create_all(bind=engine)
# app/database.py
import os
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker, declarative_base
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")

engine = create_engine(
    DATABASE_URL,
    connect_args={"check_same_thread": False}  # SQLite requirement
)

SessionLocal = sessionmaker(
    autocommit=False,
    autoflush=False,
    bind=engine
)

Base = declarative_base()

# FastAPI dependency
def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

