document_backend/requirements.txt
# =========================
# FastAPI Backend
# =========================
fastapi
uvicorn[standard]

# =========================
# Database / ORM
# =========================
sqlalchemy
pydantic
pydantic-settings

# =========================
# File Uploads
# =========================
python-multipart

# =========================
# AI & LangChain
# =========================
langchain-ollama
langchain
langchain-community
langchain-text-splitters
langgraph
python-dotenv